requirements.txt
slack_bolt>=1.18.0
slack_sdk>=3.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
paramiko>=3.4.0
requests>=2.31.0
pymysql>=1.1.0
